requirements.txt
absl-py==2.1.0
altair==5.4.1
anyio==4.4.0
astor==0.8.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
chefboost==0.0.18
click==8.1.7
cmake==3.30.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
deepface==0.0.92
dlib==19.24.6
exceptiongroup==1.2.2
facenet-pytorch==2.5.2
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
fst-pso==1.8.1
FuzzyTM==2.0.9
gast==0.6.0
gdown==5.2.0
gensim==4.3.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.65.4
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.39.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
miniful==0.0.6
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
mxnet==1.6.0
namex==0.0.8
narwhals==1.5.5
networkx==3.3
numba==0.56.4
numpy==1.23.5
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==1.5.3
pillow==10.2.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydeck==0.9.1
pyFUME==0.3.4
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.10.1
seaborn==0.13.2
simpful==2.12.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.38.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
timm==1.0.8
toml==0.10.2
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
ultralytics==8.2.82
ultralytics-thop==2.0.5
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
